packages/api/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile --output-file=/app/requirements.txt /app/requirements.in
#
annotated-doc==0.0.4
    # via fastapi
annotated-types==0.7.0
    # via pydantic
anyio==4.13.0
    # via
    #   httpx
    #   starlette
    #   watchfiles
bcrypt==3.2.2
    # via
    #   -r /app/requirements.in
    #   passlib
certifi==2026.4.22
    # via
    #   httpcore
    #   httpx
    #   requests
cffi==2.0.0
    # via
    #   bcrypt
    #   cryptography
charset-normalizer==3.4.7
    # via
    #   pdfminer-six
    #   requests
click==8.3.3
    # via uvicorn
cryptography==47.0.0
    # via
    #   -r /app/requirements.in
    #   pdfminer-six
    #   python-jose
dnspython==2.8.0
    # via email-validator
ecdsa==0.19.2
    # via python-jose
email-validator==2.3.0
    # via -r /app/requirements.in
et-xmlfile==2.0.0
    # via openpyxl
fastapi==0.136.1
    # via -r /app/requirements.in
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.9
    # via httpx
httptools==0.7.1
    # via uvicorn
httpx==0.28.1
    # via -r /app/requirements.in
idna==3.13
    # via
    #   anyio
    #   email-validator
    #   httpx
    #   requests
lxml==6.1.0
    # via
    #   python-docx
    #   python-pptx
numpy==2.4.4
    # via pandas
openpyxl==3.1.5
    # via -r /app/requirements.in
packaging==26.2
    # via pytesseract
pandas==3.0.2
    # via -r /app/requirements.in
passlib[bcrypt]==1.7.4
    # via -r /app/requirements.in
pdfminer-six==20251230
    # via
    #   -r /app/requirements.in
    #   pdfplumber
pdfplumber==0.11.9
    # via -r /app/requirements.in
pillow==12.2.0
    # via
    #   -r /app/requirements.in
    #   pdfplumber
    #   pytesseract
    #   python-pptx
psycopg2-binary==2.9.12
    # via -r /app/requirements.in
pyasn1==0.6.3
    # via
    #   python-jose
    #   rsa
pycparser==3.0
    # via cffi
pydantic==2.13.3
    # via
    #   -r /app/requirements.in
    #   fastapi
    #   pydantic-settings
pydantic-core==2.46.3
    # via pydantic
pydantic-settings==2.14.0
    # via -r /app/requirements.in
pypdfium2==5.7.1
    # via pdfplumber
pytesseract==0.3.13
    # via -r /app/requirements.in
python-dateutil==2.9.0.post0
    # via pandas
python-docx==1.2.0
    # via -r /app/requirements.in
python-dotenv==1.2.2
    # via
    #   -r /app/requirements.in
    #   pydantic-settings
    #   uvicorn
python-jose[cryptography]==3.5.0
    # via -r /app/requirements.in
python-multipart==0.0.27
    # via -r /app/requirements.in
python-pptx==1.0.2
    # via -r /app/requirements.in
pyyaml==6.0.3
    # via
    #   -r /app/requirements.in
    #   uvicorn
requests==2.33.1
    # via
    #   -r /app/requirements.in
    #   resend
resend==2.29.0
    # via -r /app/requirements.in
rsa==4.9.1
    # via python-jose
six==1.17.0
    # via
    #   ecdsa
    #   python-dateutil
starlette==1.0.0
    # via fastapi
tabulate==0.10.0
    # via -r /app/requirements.in
typing-extensions==4.15.0
    # via
    #   anyio
    #   fastapi
    #   pydantic
    #   pydantic-core
    #   python-docx
    #   python-pptx
    #   resend
    #   starlette
    #   typing-inspection
typing-inspection==0.4.2
    # via
    #   fastapi
    #   pydantic
    #   pydantic-settings
urllib3==2.6.3
    # via requests
uvicorn[standard]==0.46.0
    # via -r /app/requirements.in
uvloop==0.22.1
    # via uvicorn
watchfiles==1.1.1
    # via uvicorn
websockets==16.0
    # via uvicorn
xlsxwriter==3.2.9
    # via python-pptx

pytest==8.0.0
pytest-asyncio==0.23.5
